+++ packages/smarthr-ui/src/components/AppHeader/components/mobile/NavigationContext.tsx
@@ -1,0 +1,9 @@
+import React, { Dispatch, createContext } from 'react'
+
+import { Navigation, NavigationGroup } from '../../types'
+
+export const NavigationContext = createContext({
+  navigations: [] as Navigation[],
+  selectedNavigationGroup: null as NavigationGroup | null,
+  setSelectedNavigationGroup: (() => {}) as Dispatch<React.SetStateAction<NavigationGroup | null>>,
+})