+++ packages/smarthr-ui/src/components/ComboBox/ListBoxItemButton.tsx
@@ -66,7 +66,6 @@ const ListBoxItemButton = <T,>({
   )
 
   return isNew ? (
-    // eslint-disable-next-line smarthr/a11y-clickable-element-has-text
     <button
       type="button"
       key={option.id}