+++ packages/smarthr-ui/src/components/DatePicker/stories/VRTDatePicker.stories.tsx
@@ -11,7 +11,7 @@ import { DatePicker } from '../DatePicker'
 import type { Meta, StoryObj } from '@storybook/react'
 
 export default {
-  title: 'Forms（フォーム）/DatePicker/VRT',
+  title: 'Forms（フォーム）/DatePicker（非推奨）/VRT',
   component: DatePicker,
   render: (args) => {
     const value = '2024/11/06'
@@ -55,7 +55,7 @@ export default {
   parameters: {
     chromatic: { disableSnapshot: false },
   },
-  tags: ['!autodocs'],
+  tags: ['!autodocs', 'skip-test-runner'],
 } satisfies Meta<typeof DatePicker>
 
 export const VRT = {}