+++ packages/smarthr-ui/src/components/Fieldset/Fieldset.tsx
@@ -2,6 +2,6 @@ import React, { ComponentProps } from 'react'
 
 import { ActualFormControl } from '../FormControl/FormControl'
 
-export const Fieldset: React.FC<Omit<ComponentProps<typeof ActualFormControl>, 'as'>> = (props) => (
-  <ActualFormControl {...props} as="fieldset" />
-)
+export const Fieldset: React.FC<
+  Omit<ComponentProps<typeof ActualFormControl>, 'as' | 'htmlFor' | 'labelId'>
+> = (props) => <ActualFormControl {...props} as="fieldset" />