+++ packages/smarthr-ui/src/components/SideNav/SideNavItemButton.tsx
@@ -59,19 +59,24 @@ export const SideNavItemButton: FC<Props & ElementProps> = ({
   id,
   title,
   prefix,
-  isSelected = false,
-  size = 'default',
+  isSelected,
+  size,
   onClick,
   children,
 }) => {
-  const handleClick = onClick
-    ? (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => onClick(e, id)
-    : undefined
+  const handleClick = useMemo(
+    () =>
+      onClick
+        ? (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => onClick(e, e.currentTarget.value)
+        : undefined,
+    [onClick],
+  )
 
   const { wrapperStyle, buttonStyle, buttonInnerStyle } = useMemo(() => {
     const { wrapper, button, buttonInner } = sideNavItem()
+
     return {
-      wrapperStyle: wrapper({ selected: isSelected }),
+      wrapperStyle: wrapper({ selected: !!isSelected }),
       buttonStyle: button({ size }),
       buttonInnerStyle: buttonInner(),
     }
@@ -79,7 +84,7 @@ export const SideNavItemButton: FC<Props & ElementProps> = ({
 
   return (
     <li className={wrapperStyle}>
-      <UnstyledButton onClick={handleClick} className={buttonStyle}>
+      <UnstyledButton className={buttonStyle} onClick={handleClick} value={id}>
         <Cluster inline align="center" as="span">
           {prefix}
           <span className={buttonInnerStyle}>{children ?? title}</span>